+++ Desegundos_aminutos.py
@@ -2,7 +2,7 @@ print("Algoritmo que imprima que dada cantidad de segundo indique cuantos minuto
 
 segundos=int(input("Digita la cantidad de segundos que deseas convertir a minutos: "))
 
-minutos=segundos/60
+minutos=int(segundos/60)
 
 print("En " + str(segundos) + " segundos hay " + str(minutos) + " minutos")
 print("Final del programa")
